Improved map builder and added error handler
The map builder now returns Results if something is wrong with the
length of the strings.  The error it produces will be converted into
a general GameError which is a summary of all errors of the whole game.
The GameError can decide the severity of the error which also defines if
an error should just produce a log entry or if it is so critical that the
game must be stopped.